auth: Improve credentials management
Currently, if the users didn't log in using 'az', they will need to re-auth
on each command run. We improve this by using 'tokenCache'
to save the credentials locally. The implementation uses 'cache.ExportReplace' allowing
saving 'RefrestToken' as well so user stays logged in for longer intervals as well.
'az login' is following the same approach as well.