run-command: Truncate output tail instead of head by default
We want to truncate the tail by default because most of commands used
for debugging print a column header which is necessary to understand the
output. In addition, if the output is too long, those tools usually
provide mechanisms for filtering. Notice it is the opposite behaviour of
the Azure CLI.